requirements.txt
Django==1.11
djangorestframework==3.6.4
pytz==2018.9
